% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChisqTest_expect.R
\name{ChisqTest_expect}
\alias{ChisqTest_expect}
\title{Chi-square test}
\usage{
ChisqTest_expect(
  vec_r_events_control,
  vec_n_sample_size_control,
  vec_n_sample_size_treated,
  OR_hat,
  n,
  alpha,
  vec_pos
)
}
\arguments{
\item{vec_r_events_control}{an ordered vector of number of events in the untreated group of constituent studies from a meta-analysis.}

\item{vec_n_sample_size_control}{an ordered vector of the number of participants in the untreated group.}

\item{vec_n_sample_size_treated}{an ordered vector of the number of participants in the treated group.}

\item{OR_hat}{Summary odds ratio from meta-analysis}

\item{n}{Number of iterations used to generate constituent study power; suggest use 10,000.}

\item{alpha}{Type-1 error rate.}

\item{vec_pos}{Vector of positive results from constituent studies, returned by test.n.treated function.}
}
\value{
Vector of p-values for difference between observed and expected number of positive studies from meta-analysis, along with vector of expected values.
}
\description{
for a significant difference between observed and expected number of positive studies, not for end user.
}
\details{
This is an internal function, not for end user.
}
