% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.n.treated.R
\name{test.n.treated}
\alias{test.n.treated}
\title{Observed number of positive by testing for significance from observed findings at given alpha level (not intended for end user).}
\usage{
test.n.treated(
  vec_r_events_control,
  vec_r_events_treated,
  vec_n_sample_size_control,
  vec_n_sample_size_treated,
  alpha
)
}
\arguments{
\item{vec_r_events_control}{an ordered vector of number of events in the untreated group of constituent studies from a meta-analysis.}

\item{vec_r_events_treated}{an ordered vector of number of events in the treated group of constituent studies from a meta-analysis.}

\item{vec_n_sample_size_control}{an ordered vector of the number of participants in the untreated group.}

\item{vec_n_sample_size_treated}{an ordered vector of the number of participants in the treated group.}

\item{alpha}{Type-1 error rate.}
}
\value{
Vector of results, 1 if positive, 0 if negative.
}
\description{
by testing for significance from observed findings at given alpha level (not intended for end user).
}
\details{
Not for end user.
}
