% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.one.treated.R
\name{test.one.treated}
\alias{test.one.treated}
\title{Is one constituent study observed significant, in favour of treatment, at a given alpha level? (Not intended for end user)}
\usage{
test.one.treated(
  r_events_control,
  r_events_treated,
  n_sample_size_control,
  n_sample_size_treated,
  alpha = 0.05
)
}
\arguments{
\item{r_events_control}{Number of events in untreated}

\item{r_events_treated}{Number of events in treated group}

\item{n_sample_size_control}{Sample size in untreated group}

\item{n_sample_size_treated}{Sample size in treated group}

\item{alpha}{Type-1 error rate}
}
\value{
Number, 1 if positive, 0 if negative
}
\description{
Is one constituent study observed significant, in favour of treatment, at a given alpha level? (Not intended for end user)
}
