\docType{data}
\name{BMort}
\alias{BMort}
\title{BMort}
\format{Brugts meta-analysis study of effect of statins on overall mortality (BMJ 2009).}
\source{
  Brugts JJ, Yetgin T, Hoeks SE, Gotto AM, Shepherd J,
  Westendorp RGJ, et al. The benefits of statins in people
  without established cardiovascular disease but with
  cardiovascular risk factors: meta-analysis of randomised
  controlled trials. BMJ 2009;338
}
\description{
  BMort
}
\keyword{datasets}

