% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightrMeta.R
\name{weightrMeta}
\alias{weightrMeta}
\title{weightr}
\usage{
weightrMeta(meta)
}
\arguments{
\item{meta}{a meta-analysis object (e.g. `rma.uni`)}
}
\value{
a data frame with infomation passed from weightr containing the meta-analysis
}
\description{
`weightrMeta()` Connects the weightr package to the Metaverse.
}
\examples{
\dontrun{
library(metafor)
library(broom)
ma_model_molloy2014 <- metafor::dat.molloy2014


ma <- ma_model_molloy2014 \%>\%
 escalc(
   measure = "ZCOR",
   ni = ni,
   ri = ri,
   data = .) \%>\%
 group_by(design) \%>\%
 meta_analysis(yi = yi, vi = vi, slab = authors, exponentiate = TRUE)

weightrMeta(ma)
}
}
