% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightrMeta.R
\name{weightrMeta}
\alias{weightrMeta}
\title{weightr}
\usage{
weightrMeta(meta, ...)
}
\value{
a `tbl` with a `listcol`, `meta`, containing the meta-analysis
}
\description{
`weightrMeta()` Connects the weightr package to the Metaverse
}
\details{
#@param meta a meta-analysis object (e.g. `rma.uni`)
}
\examples{
\dontrun{
library(metafor)
library(broom)

ma <- rma(yi = lnes, sei = selnes, slab = study_name, data = iud_cxca)

ma \%>\%
  tidy() \%>\%
  weightrMeta(ma)
}
}
